feat : 게시글 길이 및 Navbar 갱신 문제 수정
