+++ src/app/Dashboard/page.tsx
@@ -5,5 +5,5 @@ import { getTags } from "@/lib/server/tags";
 export default async function DashboardPage() {
   const tags = await getTags(); // 서버에서 태그 가져오기
 
-  return <BoardPage initialTags={tags} />;
+  return <BoardPage />;
 }
