+++ src/app/login/createMember/page.tsx
@@ -80,7 +80,8 @@ const CreateMember = () => {
       }
 
       console.log("세션 종료");
-      router.push("/"); // 변경사항: Next.js 방식으로 홈으로 리다이렉트
+      //router.push("/"); 1변경안 :  Next.js 방식으로 홈으로 리다이렉트
+      window.location.href = "/"; // 2변경안 : 로그아웃 후 메인페이지가 강제로 리렌더링하도록 설정
     } catch (err) {
       console.error(err);
       setError("회원가입 처리 중 오류가 발생했습니다.");
@@ -104,92 +105,103 @@ const CreateMember = () => {
     );
 
   return (
-    <div className="flex flex-col min-h-screen justify-center py-12 px-4 sm:px-6 lg:px-8 bg-gray-50">
-      <div className="sm:mx-auto sm:w-full sm:max-w-md">
-        {/* 로고 및 타이틀 */}
-        <h2 className="text-center text-3xl font-bold text-blue-600 mb-2">
-          캠퍼스 커넥트
-        </h2>
-        <h3 className="text-center text-xl text-gray-600 mb-6">
-          {name}님 반갑습니다. 회원가입을 진행해주세요.
-        </h3>
-      </div>
+    <div
+      id="CreateMemberbg"
+      className="flex w-full min-h-screen justify-center items-center bg-gray-50"
+    >
+      {/* 로그인 폼 박스 */}
+      <div
+        id="CreateMemberForm"
+        className="flex flex-col min-w-[390px] h-auto justify-center py-12 px-4 sm:px-6 lg:px-8 bg-white"
+      >
+        {/* 제목 박스 */}
+        <div className="whitespace-nowrap sm:mx-auto sm:w-full sm:max-w-md">
+          <h2 className="text-center text-3xl font-bold text-blue-600 mb-2">
+            HongikMentor
+          </h2>
+          <h3 className="text-center text-xl text-gray-600 mb-6">
+            {name}님 반갑습니다. 회원가입을 진행해주세요.
+          </h3>
+        </div>
 
-      <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
-        <div className="bg-white py-8 px-4 shadow-lg rounded-lg sm:px-10">
-          <form className="space-y-6" onSubmit={handleSubmit}>
-            <div>
-              <label
-                htmlFor="name"
-                className="block text-sm font-medium text-gray-700"
-              >
-                이름
-              </label>
-              <div className="mt-1">
-                <input
-                  id="name"
-                  name="name"
-                  type="text"
-                  required
-                  value={name}
-                  onChange={(e) => setName(e.target.value)}
-                  placeholder="이름을 입력하세요"
-                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                />
+        {/* 로그인 폼 */}
+
+        <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
+          <div className="bg-white py-8 px-4 shadow-lg rounded-lg sm:px-10">
+            <form className="space-y-6" onSubmit={handleSubmit}>
+              <div>
+                <label
+                  htmlFor="name"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  이름
+                </label>
+                <div className="mt-1">
+                  <input
+                    id="name"
+                    name="name"
+                    type="text"
+                    required
+                    value={name}
+                    onChange={(e) => setName(e.target.value)}
+                    placeholder="이름을 입력하세요"
+                    className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+                  />
+                </div>
               </div>
-            </div>
 
-            <div>
-              <label
-                htmlFor="major"
-                className="block text-sm font-medium text-gray-700"
-              >
-                전공
-              </label>
-              <div className="mt-1">
-                <input
-                  id="major"
-                  name="major"
-                  type="text"
-                  required
-                  value={major}
-                  onChange={(e) => setMajor(e.target.value)}
-                  placeholder="전공을 입력하세요"
-                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                />
+              <div>
+                <label
+                  htmlFor="major"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  전공
+                </label>
+                <div className="mt-1">
+                  <input
+                    id="major"
+                    name="major"
+                    type="text"
+                    required
+                    value={major}
+                    onChange={(e) => setMajor(e.target.value)}
+                    placeholder="전공을 입력하세요"
+                    className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+                  />
+                </div>
               </div>
-            </div>
 
-            <div>
-              <label
-                htmlFor="graduationYear"
-                className="block text-sm font-medium text-gray-700"
-              >
-                졸업년도
-              </label>
-              <div className="mt-1">
-                <input
-                  id="graduationYear"
-                  name="graduationYear"
-                  type="number"
-                  required
-                  value={graduationYear}
-                  onChange={(e) => setGraduationYear(e.target.value)}
-                  placeholder="예: 2025"
-                  className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                />
+              <div>
+                <label
+                  htmlFor="graduationYear"
+                  className="block text-sm font-medium text-gray-700"
+                >
+                  졸업년도
+                </label>
+                <div className="mt-1">
+                  <input
+                    id="graduationYear"
+                    name="graduationYear"
+                    type="number"
+                    required
+                    value={graduationYear}
+                    onChange={(e) => setGraduationYear(e.target.value)}
+                    placeholder="예: 2025"
+                    className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500"
+                  />
+                </div>
               </div>
-            </div>
 
-            <div>
-              <button
-                type="submit"
-                className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 transition-colors"
-              >
-                가입하기
-              </button>
-            </div>
-          </form>
+              <div>
+                <button
+                  type="submit"
+                  className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 transition-colors"
+                >
+                  가입하기
+                </button>
+              </div>
+            </form>
+          </div>
         </div>
       </div>
     </div>
